update
one enemy will now spawn to be able to see if collision works.

gave said enemy a starting position in the middle of the screen and a slow base speed for testing

increased the speed of the bullet and the rate of fire so that the player shoots faster and more making the player capable of killing the enemies when more are on screen(will need further optimization).